active-settings is a unique application
Use KDBusService to register and handle the unique application mode.
This is only done once we have loaded the UI, as we may want to call
into it from the outside. Doing it this late may save some headache in
race conditions when the app starts twice.
That said, we may want to have it *quit* much earlier...