Load the specified module in the "main application"
Recycle command-line arguments, and use them to update the view "from
the outside". This means that a second instance of active-settings,
loading a module with the command line option -m and --module will cause
the specified module to be loaded in the existing application window.

It also means that we centralize all settings access into one app, which
is a sensible pattern for a "system configuration" thing.