[kaccounts] Add carddav contacts sync plugin
This is based on Jolla's Buteo carddav sync plugin with the following
changes:

* stripped away all the Nemo/Jolla specific stuff
* stripped away the QSqlLite stuff
* replaced any QContact usage with KContacts::Addressee
* replaced the custom vCard converter with KContacts::VCardConverter
* replaced custom LOG_* with qDebug/qWarning
* added CMakeLists.txt

Still quite some portion of the code is left commented out and needs
adapting to our needs. So far it downloads the contacts and stores them
in vcards, but it also has code for sending back changes.