optionally use apply buttons instead of instant apply
This change starts respecting the "supportsInstantApply" property of a
KCM. If this is set to false, we will display a toolbar with Apply,
Reset and Default buttons as requested by the KCM and use those to
trigger the respective actions instead of instantly applying. Help is
not shown, on mobile, we expect the UI to be self-explanatory.

Signed-off-by: Sebastian Kŭgler <sebas@kde.org>